pyproject.toml
[tool.poetry]
name = "stf-nomic-game"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.scripts]
start = "uvicorn app.main:app --reload"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
bcrypt = "3.2.0"
websockets = "^12.0"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"

[tool.poetry.dev-dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
